+++ InputManager.h
@@ -17,12 +17,12 @@ private:
     void addKeyboardMapping(sf::Keyboard::Key input_k);
     void addMouseMapping(sf::Mouse::Button input_b);
     void parseEvents(sf::Event event);
-    sf::Vector2f getMousePosition();
+    sf::Vector2i getMousePosition();
 
     std::map<sf::Keyboard::Key, bool> keyboard_map;
     std::map<sf::Mouse::Button, bool> mouse_map;
     sf::RenderWindow* window;
-    sf::Vector2f mousePos;
+    sf::Vector2i mousePos;
 };
 
 #endif
